refactor(api): SCALE-SEAM ⑩ (targeted) — /finance out of client.ts; pin 3,748 -> 3,726
CI caught this, not me: `test_file_sizes` failed with
`apps/web/src/api/client.ts=3757 > 3748`. Giving `financeReconcile` the response
shape it actually returns — instead of three `unknown[]` buckets that sent its
first caller to read `fin_ingest.reconcile` — cost nine lines, and client.ts sat
at exactly the pin I set in ⑨.

A GENERAL SCALE-SEAM ⑩ WAS DROPPED, AND THIS IS NOT THAT. Measuring the residue
showed 519 methods across 191 route groups averaging 2.7 each — no coherent
domain left, and the largest remaining group would have moved 55 of 3,748 lines
while moving the reachability ceiling by zero. That plan died on those numbers.

This one the ratchet asked for by name. The file's own comment says the friction
is the point and should force "should this live in a domain module instead?" —
and for a four-method route group that now has a panel of its own, the answer is
yes. So the pin goes DOWN 22 rather than up 9.

PROOF NOTHING WAS DROPPED, from the ratchet's own reader: 705 methods reachable
with the extraction, the same number as without it. A move is invisible to that
reader; a loss is not.

Suite 1416 / 128; tsc, lint, size gate green.
