+++ apps/web/src/api/client.ts
@@ -5,6 +5,7 @@ import { withAuthoring } from "./authoring";
 import { withRoutines } from "./routines";
 import { withContracts } from "./contracts";
 import { withDesignOptions } from "./designOptions";
+import { withFinance } from "./finance";
 import { withLibrary } from "./library";
 import { HttpCore, type LiveStream } from "./httpCore";
 import { withModel } from "./model";
@@ -37,7 +38,7 @@ import type {
 
 // Transport (baseUrl, token, json/_pdfPost/url/health) lives in HttpCore; ApiClient adds the typed
 // domain methods below. Every `api.method()` call site is unchanged by the split.
-export class ApiClient extends withContracts(withAuth(withProforma(withDesignOptions(withRoutines(withProcurement(withEstimate(withModules(withModel(withSchedule(withLibrary(withAuthoring(HttpCore)))))))))))) {
+export class ApiClient extends withFinance(withContracts(withAuth(withProforma(withDesignOptions(withRoutines(withProcurement(withEstimate(withModules(withModel(withSchedule(withLibrary(withAuthoring(HttpCore))))))))))))) {
   /** Admin: integration settings (AI / email / SSO). Secret values are never returned. */
   integrations() {
     return this.json<{ groups: IntegrationGroup[] }>("/settings/integrations");
@@ -2119,38 +2120,6 @@ export class ApiClient extends withContracts(withAuth(withProforma(withDesignOpt
       `/projects/${pid}/propmap/plan`, { method: "POST", body: JSON.stringify({ rules }) });
   }
 
-  // real-estate development finance (Proforma)
-  /** FIN-GOV — the project's locked reporting period (books closed through lock_date, or null). */
-  financeLock(pid: string) {
-    return this.json<{ lock_date: string | null; set_by?: string; set_at?: string; note?: string }>(
-      `/projects/${pid}/finance/lock`);
-  }
-  setFinanceLock(pid: string, lockDate: string | null, note?: string) {
-    return this.json<{ lock_date: string | null; set_by: string; note: string }>(
-      `/projects/${pid}/finance/lock`,
-      { method: "PUT", body: JSON.stringify({ lock_date: lockDate, note: note ?? "" }) });
-  }
-  /** FIN-INGEST — budget ↔ actuals two-way reconciliation on the cost-code spine.
-   *
-   *  The three buckets were typed `unknown[]`, which meant the first caller had to go read
-   *  `fin_ingest.reconcile` to render a row — so the type is now the shape the server actually
-   *  returns. All three carry the same slim row; they differ only in which side had a value, and
-   *  they are deliberately NOT netted against each other.
-   */
-  financeReconcile(pid: string) {
-    type Row = { cost_code: string | null; budget: number | null; committed: number | null;
-      actual: number | null; variance: number | null };
-    return this.json<{ matched: Row[]; budget_only: Row[]; actuals_only: Row[];
-      uncoded: { module: string; ref: string; amount: number; vendor?: string }[];
-      counts: { matched: number; budget_only: number; actuals_only: number; uncoded: number };
-      fully_reconciled: boolean }>(
-      `/projects/${pid}/finance/reconcile`);
-  }
-  /** FIN-INGEST — import lineage: the project's audit-logged import batches, newest first. */
-  financeImports(pid: string) {
-    return this.json<{ ts: string | null; actor: string | null; module: string; filename: string;
-      imported: number; error_count: number }[]>(`/projects/${pid}/finance/imports`);
-  }
   /** CRE-HOLDSELL — hold vs sell: incremental hold-year IRRs against the proceeds declined today. */
   holdSell(pid: string, inputs: unknown, hurdleRate = 0.12, maxYears = 10) {
     return this.json<{ computable: boolean; reason?: string;
